+++ apps/website/app/ui/docs/components/menu/layout.tsx
@@ -1,13 +1,20 @@
+
 import React from 'react';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
-  title:
-    'gluestack-ui Menu Component | React Native Menu Installation, Usage & API',
-  description:
-    'Build a user-friendly interface with gluestack-ui menu component in React & React Native, designed for easy navigation and accessibility.',
-};
+  title: "gluestack-ui Menu Component | React Native Menu Installation, Usage & API",
+  description: "Build a user-friendly interface with gluestack-ui menu component in React & React Native, designed for easy navigation and accessibility."
+}
 
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
+export default function Layout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+    </>
+  );
 }