+++ apps/website/app/ui/docs/home/getting-started/gluestack-ui-nativewind-utils/layout.tsx
@@ -1,3 +1,4 @@
+
 import React from 'react';
 import { Metadata } from 'next';
 
@@ -6,6 +7,14 @@ export const metadata: Metadata = {
   description: "gluestack-ui/nativewind-utils provides a collection of utility functions for your gluestack-ui components. A must-have NativeWind UI library for faster and smoother UI development."
 }
 
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return <>{children}</>;
+export default function Layout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <>
+      {children}
+    </>
+  );
 }