+++ packages/gluestack-ui/templates/nextjs/next.config.mjs
@@ -4,6 +4,9 @@ const nextConfig = {
   typescript: {
     ignoreBuildErrors: true,
   },
+  eslint: {
+    ignoreDuringBuilds: true,
+  },
   webpack: (config, { isServer }) => {
     // Configure webpack for standard builds
     config.resolve.alias = {