+++ packages/gluestack-ui/CHANGELOG.md
@@ -1,3 +1,87 @@
 ## 0.4.9-alpha.0 (2023-07-21)
 
+## 0.1.33
+
+### Patch Changes
+
+- chore: removed unneccassary console log
+
+## 0.1.24
+
+### Patch Changes
+
+- chore: removed reanimated from expo deps
+
+## 0.1.32
+
+### Patch Changes
+
+- chore: updated babel config
+
+## 0.1.31
+
+### Patch Changes
+
+- chore: testing
+
+## 0.1.30
+
+### Patch Changes
+
+- chore: test
+
+## 0.1.29
+
+### Patch Changes
+
+- chore: testing
+
+## 0.1.28
+
+### Patch Changes
+
+- chore: minor config changes
+
+## 0.1.27
+
+### Patch Changes
+
+- fix: updated reanimated init
+
+## 0.1.26
+
+### Patch Changes
+
+- fix: react-native-reanimated dependency issue
+
+## 0.1.25
+
+### Patch Changes
+
+- fix: updated postcss logic
+
+## 0.1.23
+
+### Patch Changes
+
+- chore: fixed postcss issue
+
+## 0.1.21
+
+### Patch Changes
+
+- chore: fixes testing
+
+## 0.1.17
+
+### Patch Changes
+
+- chore: testinf
+
+## 0.1.13
+
+### Patch Changes
+
+- fix: removed unneccessary gluestack-ui-config.ts init
+
 - Testing